requirements.txt
Flask==2.0.3
uvicorn==0.29.0
paho-mqtt==1.6.1
grpcio==1.64.0
grpcio-tools==1.64.0
pytest==8.1.1
pytest-asyncio==0.23.6
requests==2.31.0
python-dotenv==1.0.1
loguru==0.7.2


tweepy~=4.15.0
protobuf~=5.29.3
pdfplumber~=0.11.5
redis~=5.2.1